benchmarks/llm-provider/requirements.txt
openai>=1.30.0
anthropic>=0.28.0
python-dotenv>=1.0.0
rich>=13.7.0
